Add UffTable load and lookup methods, fix uff table tests.
Removed one test that checks the length of returned row. Since we will always return one single row and this is already checked in lookup method there is no need for an additional test.

Added docstring for exceptions.

Added lookupo and load methods for uff table read table and return rows for given element, coordination number and angle. The exceptions are raised in lookup method for different cases.